fix decimal angle y

+++ src/java/ff/server/SerialPortCommServer.java
@@ -40,7 +40,7 @@ public class SerialPortCommServer {
     // key为ip,value为true或false，当value为false时，当前正在巡航的云台。当前正在巡航的云台，不会重复发送巡航右转命令。以减少命令发送量。
     private static Map<String, Boolean> isCruising = new ConcurrentHashMap<String, Boolean>();
     // 为巡航，比如削苹果皮等准备。其中Key为ip.value为预置的Y角度，只有达到此角度时才接受其它命令。
-    private static Map<String, Integer> isCruisingPresetAngleY = new ConcurrentHashMap<String, Integer>();
+    private static Map<String, Double> isCruisingPresetAngleY = new ConcurrentHashMap<String, Double>();
     //火警时，如果正在调整起火点到中心位置，则以下变量中的状态为true;
     private static Map<String, Boolean> isMovingCenterForFireAlarm = new ConcurrentHashMap<String, Boolean>();
     //火警时，当场保存的起火方位与热值，时间。由于火警地点随时变化，所以这里只记录最初的方位。Key为IP，值为：X|Y|AngleX|AngleY|MaxValue|Time
@@ -562,7 +562,7 @@ public class SerialPortCommServer {
         return isCruising;
     }
 
-    public Map<String, Integer> getIsCruisingPresetAngleY() {
+    public Map<String, Double> getIsCruisingPresetAngleY() {
         return isCruisingPresetAngleY;
     }
